requirements.txt
pandas
streamlit
psycopg2
python-dotenv
matplotlib
seaborn
plotly
statsmodels
apache-airflow
asyncpg
sqlalchemy
apache-airflow-providers-postgres